fix[qboot][update]: harden firmware release and stream validation
validate firmware package and destination sizes before release/update
verify destination image before skipping already signed packages
harden stream decompression progress and AES block-aligned reads
guard update-manager callbacks and reject invalid reentrant state transitions
add host-side hooks for storage/sign fault injection

+++ algorithm/qboot_aes.c
@@ -32,7 +32,8 @@ static tiny_aes_context qbt_aes_ctx;
  */
 void qbt_aes_decrypt_init(void)
 {
-    int len = strlen(QBOOT_AES_IV);
+    size_t len = strlen(QBOOT_AES_IV);
+
     if (len > sizeof(qbt_aes_iv))
     {
         len = sizeof(qbt_aes_iv);
